Pessoa/Back Pessoa/requirements.txt
mysql-connector-python = 8.0.31
mysqlclient = 2.1.1
Flask = 2.2.2